Add flash decode runner to benchmarks
